% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secret.R
\name{gsecret_auth}
\alias{gsecret_auth}
\title{Authorize gsecretR}
\usage{
gsecret_auth(
  email = gargle::gargle_oauth_email(),
  path = NULL,
  scopes = "https://www.googleapis.com/auth/cloud-platform",
  cache = gargle::gargle_oauth_cache(),
  use_oob = gargle::gargle_oob_default(),
  token = NULL
)
}
\arguments{
\item{email}{Optional, your google identity}

\item{path}{path to GCP JSON key file for a service account}

\item{scopes}{required scopes.}

\item{cache}{Specifies the OAuth token cache}

\item{use_oob}{Should we use out-of-band authentication}

\item{token}{Pass in a token}
}
\value{
A token
}
\description{
Authorize gsecretR to access or create secrets stored in Secret Manager on Google Cloud Platform.
This function is based on [bigrquery::bq_auth()] and also wraps [gargle::token_fetch()].
}
