% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secret.R
\name{list_secrets}
\alias{list_secrets}
\title{List secrets within a project}
\usage{
list_secrets(project_id)
}
\arguments{
\item{project_id}{the project containing the secrets.}
}
\value{
a data frame containing the secrets
}
\description{
List secrets within a project
}
