% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secret.R
\name{get_secret}
\alias{get_secret}
\alias{set_secret}
\title{Set a secret on Google Cloud}
\usage{
get_secret(project_id, secret_id, version_id = "latest")

set_secret(project_id, secret_id, secret)
}
\arguments{
\item{project_id}{the project storing (and charged for) the secret}

\item{secret_id}{the id of the secret}

\item{version_id}{version id retrieved by get_secret}

\item{secret}{a string secret}
}
\description{
\code{set_secret()} assigns a secret to an id in project.
\code{get_secret()} retrieves the secret with id from a project
by default, \code{get_secret()} returns the latest version of a
secret, but if you know the secret, you can select it by the version.
}
