% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secret.R
\name{gsecret_auth_config}
\alias{gsecret_auth_config}
\title{Configure the authentication}
\usage{
gsecret_auth_config(app)
}
\arguments{
\item{app}{\itemize{
\item an oauth authenicator app containing a client id/secret
see \code{\link[gargle:gargle_app]{gargle::gargle_app()}}.
}}
}
\description{
Configure the authentication
}
